Implement basic image grabbing
Uses a proxy.js file wrote using express, request and path npm modules that can be found in my WebDev-Snippets repo. Images are grabbed from the reddit api from /r/EarthPorn/hot.json and are then passed through the proxy to avoid CORS and then a gradient is generated from them using the canvas. A lot of the pixel grabbing, if not all of it, was temporary from before the proxy was implemented, I will probably switch back to using a non-canvas based method so it is slightly faster.